restructure cmor config
